move build_state_override_maps to state_override_view
